refactor: remove cycle from AssignSignatory
