+++ src/components/Company/AssignSignatory/CreateSignatory/CreateSignatoryForm.tsx
@@ -2,7 +2,7 @@ import { useFormContext } from 'react-hook-form'
 import * as v from 'valibot'
 import { useTranslation } from 'react-i18next'
 import { CalendarDate } from '@internationalized/date'
-import { useCreateSignatory } from './CreateSignatory'
+import { useCreateSignatory } from './useCreateSignatory'
 import { TextInputField, Grid, Flex, SelectField } from '@/components/Common'
 import { DatePicker } from '@/components/Common/Inputs/DatePicker'
 import { nameValidation, zipValidation, SSN_REGEX, phoneValidation } from '@/helpers/validations'
