+++ src/components/Company/AssignSignatory/CreateSignatory/index.ts
@@ -1,5 +1,2 @@
-export {
-  CreateSignatory,
-  useCreateSignatory,
-  type CreateSignatoryDefaultValues,
-} from './CreateSignatory'
+export { CreateSignatory } from './CreateSignatory'
+export { useCreateSignatory, type CreateSignatoryDefaultValues } from './useCreateSignatory'
